# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NicoShop API",
    "version": "1.0.0",
    "description": "API REST de l'e-commerce NicoShop"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Serveur local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "product_id": { "type": "integer" },
          "product_name": { "type": "string" },
          "quantity": { "type": "integer" },
          "price": { "type": "number", "format": "float" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "user_id": { "type": "integer" },
          "customer_name": { "type": "string" },
          "customer_email": { "type": "string", "format": "email" },
          "total": { "type": "number", "format": "float" },
          "status": { "type": "string", "enum": ["pending", "completed", "shipped", "cancelled"] },
          "payment_method": { "type": "string" },
          "shipping_address": { "type": "string" },
          "shipping_region": { "type": "string" },
          "delivery_fee": { "type": "number", "format": "float" },
          "currency": { "type": "string", "enum": ["FCFA", "EUR", "USD"] },
          "created_at": { "type": "string", "format": "date-time" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/OrderItem" } }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "tags": ["Commandes"],
        "summary": "Toutes les commandes (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Liste des commandes", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } } } } },
          "403": { "description": "Accès refusé" }
        }
      },
      "post": {
        "tags": ["Commandes"],
        "summary": "Créer une commande",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items", "total"],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["id", "quantity", "price"],
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "Produit" },
                        "quantity": { "type": "integer", "example": 2 },
                        "price": { "type": "number", "example": 29.99 }
                      }
                    }
                  },
                  "total": { "type": "number", "minimum": 0, "example": 59.98 },
                  "shipping_address": { "type": "string", "example": "123 Rue Example" },
                  "shipping_region": { "type": "string", "example": "Douala" },
                  "delivery_fee": { "type": "number", "minimum": 0, "example": 5.00 },
                  "payment_method": { "type": "string", "example": "OM" },
                  "currency": { "type": "string", "enum": ["FCFA", "EUR", "USD"], "example": "FCFA" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Commande créée", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
          "400": { "description": "Données invalides" }
        }
      }
    },
    "/api/orders/user/{userId}": {
      "get": {
        "tags": ["Commandes"],
        "summary": "Commandes d'un utilisateur",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": { "description": "Commandes de l'utilisateur", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } } } } }
        }
      }
    },
    "/api/orders/{id}/status": {
      "put": {
        "tags": ["Commandes"],
        "summary": "Modifier le statut d'une commande (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "CMD-abc12345" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["pending", "completed", "shipped", "cancelled"], "example": "shipped" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Statut mis à jour" },
          "400": { "description": "Statut invalide" },
          "403": { "description": "Accès refusé" },
          "404": { "description": "Commande non trouvée" }
        }
      }
    },
    "/api/orders/{id}": {
      "delete": {
        "tags": ["Commandes"],
        "summary": "Supprimer une commande (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Commande supprimée" },
          "403": { "description": "Accès refusé" },
          "404": { "description": "Commande non trouvée" }
        }
      }
    },
...
  }
}